Fix: Resolve build errors across entire codebase after pulling from main
This commit fixes all compilation errors introduced by recent metadata and
architecture changes, enabling the full project to build successfully.

## Core Library Fixes

**MetadataSnapshot Constructor Updates**:
- Added missing `layerGroups` parameter to all MetadataSnapshot constructor calls
- Fixed JsonMetadataProvider, PostgresMetadataProvider, RedisMetadataProvider,
  SqlServerMetadataProvider

**Data Layer Fixes**:
- Added missing `using Honua.Server.Core.Extensions;` directives
- Fixed SqlViewSchemaDetector to cast IDbCommand to DbCommand for async operations
- Removed duplicate QueryDefinition record declaration

**Drone Models Architecture Fix**:
- Moved Drone models from Honua.MapSDK.Models.Drone to Honua.Server.Core.Models.Drone
- Fixes layer violation (MapSDK should not define Core domain models)
- Created LodGenerationResult.cs for point cloud processing
- Updated all references across MapSDK services and Host controllers

## MapSDK Fixes (44 build errors resolved)

**Blazor Component Fixes**:
- Fixed VR component onclick handlers (escaped quotes → single quotes)
- Fixed Drone component tuple type conversions (explicit byte casts)
- Fixed TerrainExample EventCallback bindings with MudSlider
- Fixed form binding with explicit Value/ValueChanged/ValueExpression pattern
- Fixed @media query escaping in CSS

## Host Project Fixes (30+ build errors resolved)

**Metadata Administration**:
- Fixed 13 MetadataSnapshot constructor calls in MetadataAdministrationEndpoints.cs
- Fixed 4 MetadataSnapshot constructor calls in RbacEndpoints.cs
- Used sed script for bulk find/replace of layerGroups parameter

**WMS/WFS/OGC Handler Fixes**:
- Added missing metadataRegistry parameter to WMS GetFeatureInfo handler
- Fixed ExpandedRasterMember to use positional constructor with [SetsRequiredMembers]
- Fixed WFS SQL view parameters type (IReadOnlyDictionary<string, string>)
- Commented out incomplete layer group collection code in OgcLandingHandlers
  (BuildLayerGroupCollectionId, BuildLayerGroupCollectionLinks,
   RenderLayerGroupCollectionHtml methods not yet implemented)

## Test Infrastructure Fixes

**MetadataSnapshotBuilder**:
- Added layerGroups field and parameter to test builder
- Fixed all test MetadataSnapshot constructor calls in:
  - WcsTests.cs, CswTests.cs, WmtsTests.cs
  - ODataEndpointTests.cs

## Build Status

✅ Core libraries: Clean build
✅ MapSDK: Clean build (0 errors)
✅ Host: Clean build (0 errors, 7 warnings)
✅ Test projects: Clean build

## Known Issues

OData tests are failing with JSON schema validation errors, which is a known
architectural limitation documented in docs/testing/OData-Testing-Architecture.md.
This requires Docker container-based testing to resolve (separate effort).

+++ src/Honua.MapSDK/Utilities/Drone/PointCloudLodGenerator.cs
@@ -1,4 +1,4 @@
-using Honua.MapSDK.Models.Drone;
+using Honua.Server.Core.Models.Drone;
 using Microsoft.Extensions.Logging;
 
 namespace Honua.MapSDK.Utilities.Drone;
@@ -236,6 +236,6 @@ public static class PointCloudClassificationColors
     {
         return StandardColors.TryGetValue(classification, out var color)
             ? color
-            : (200, 200, 200); // Default gray
+            : ((byte)200, (byte)200, (byte)200); // Default gray
     }
 }
